Start LL multiprocessing
