+++ src/main/java/frc/robot/subsystems/Limelight.java
@@ -1,7 +1,9 @@
 package frc.robot.subsystems;
 
 import java.util.Arrays;
+import java.util.concurrent.locks.ReadWriteLock;
 
+import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Pose3d;
 import edu.wpi.first.math.geometry.Rotation2d;
@@ -9,21 +11,30 @@ import edu.wpi.first.math.geometry.Rotation3d;
 import edu.wpi.first.math.geometry.Translation3d;
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableInstance;
+import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import frc.robot.LimelightHelpers;
 import frc.robot.LimelightHelpers.PoseEstimate;
+import frc.robot.constants.RobotConstants;
+import frc.robot.constants.VisionConstants;
 import frc.robot.subsystems.drivetrain.RAROdometry;
 
-public class Limelight {
-  private NetworkTable m_limelightTable;
+public class Limelight implements Runnable {
+  private final NetworkTable m_limelightTable;
+  private final ReadWriteLock m_stateLock;
+  private final LimelightType m_limelightType;
+  private VisionConstants m_visionConstants;
   private final String m_name;
 
   /**
    * Constructor
    */
-  public Limelight(String limelightName) {
+  public Limelight(String limelightName, ReadWriteLock lock, LimelightType llType) {
     m_name = limelightName;
+    m_stateLock = lock;
+    m_limelightType = llType;
+    m_visionConstants = new VisionConstants(1,100,0,100);
 
     m_limelightTable = NetworkTableInstance.getDefault().getTable(m_name);
   }
@@ -124,5 +135,90 @@ public class Limelight {
 
   public boolean getLightEnabled() {
     return m_limelightTable.getEntry("ledMode").getDouble(1.0) == 3;
+  }
+
+  Pose2d nullPose = new Pose2d();
+
+  private boolean isPoseZero(PoseEstimate estimate) {
+    return estimate.pose.equals(nullPose);
+  }
+
+  private boolean checkPose(PoseEstimate estimate) {
+    if (estimate == null) {
+      return false;
+    }
+
+    if (isPoseZero(estimate)) {
+      return false;
+    }
+
+    if (estimate.pose.getX() <= 0 || estimate.pose.getX() > RobotConstants.robotConfig.Field.k_length) {
+      return false;
+    }
+
+    if (estimate.pose.getY() <= 0 || estimate.pose.getY() > RobotConstants.robotConfig.Field.k_width) {
+      return false;
+    }
+
+    if (estimate.tagCount <= 0) {
+      return false;
+    }
+
+    if (Math.abs(RAROdometry.getInstance().getGyro().getRate()) > 720) {
+      return false;
+    }
+
+    return true;
+  }
+
+  // TODO: Add these to Constants when we're done testing them
+  private double xyStdDevCoefficient = 0.005;
+  private double thetaStdDevCoefficient = 0.01;
+  private double stdDevFactor = 0.5; // TODO: Add more!
+  private boolean useVisionRotation = true;
+
+  private void updatePoseWithStdDev(PoseEstimate estimate) {
+    double avgDistance = estimate.avgTagDist;
+    double xyStdDev = xyStdDevCoefficient
+        * Math.pow(avgDistance, 2.0)
+        / estimate.tagCount
+        * stdDevFactor
+        * (DriverStation.isAutonomous() ? m_visionConstants.autoStdDevScale : 1.0);
+
+    double thetaStdDev = useVisionRotation
+        ? thetaStdDevCoefficient
+            * Math.pow(avgDistance, 2.0)
+            / estimate.tagCount
+            * stdDevFactor
+            * (DriverStation.isAutonomous() ? m_visionConstants.autoStdDevScale : 1.0)
+        : Double.POSITIVE_INFINITY;
+
+        
+
+    RAROdometry.getInstance().addLLPose(estimate, VecBuilder.fill(xyStdDev, xyStdDev, thetaStdDev));
+
+    // try {
+    //   m_stateLock.writeLock().lock();
+
+    //   RAROdometry.getInstance().getPoseEstimator().addVisionMeasurement(
+    //       estimate.pose,
+    //       estimate.timestampSeconds,
+    //       VecBuilder.fill(xyStdDev, xyStdDev, thetaStdDev));
+    // } finally {
+    //   m_stateLock.writeLock().unlock();
+    // }
+  }
+
+  @Override
+  public void run() {
+    PoseEstimate estimate = getPoseEstimation();
+
+    if (checkPose(estimate)) {
+      updatePoseWithStdDev(estimate);
+    }
+  }
+
+  public enum LimelightType {
+    LL2P, LL3, LL4
   }
 }
