+++ src/main/java/frc/robot/subsystems/drivetrain/RAROdometry.java
@@ -10,10 +10,12 @@ import com.studica.frc.AHRS;
 import com.studica.frc.AHRS.NavXUpdateRate;
 
 import edu.wpi.first.math.VecBuilder;
+import edu.wpi.first.math.Vector;
 import edu.wpi.first.math.estimator.SwerveDrivePoseEstimator;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
+import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import frc.robot.LimelightHelpers.PoseEstimate;
@@ -22,28 +24,24 @@ import frc.robot.constants.RobotConstants;
 import frc.robot.constants.VisionConstants;
 import frc.robot.subsystems.Limelight;
 import frc.robot.subsystems.Subsystem;
+import frc.robot.subsystems.Limelight.LimelightType;
 
 public class RAROdometry extends Subsystem {
   private static RAROdometry m_instance;
 
   private final AHRS m_gyro;
   private final Limelight m_limelight;
-
   private final SwerveDrive m_swerve = SwerveDrive.getInstance();
-
-  private VisionConstants m_visionConstants;
-
   private SwerveDrivePoseEstimator m_poseEstimator;
 
   /** Lock used for odometry thread. */
   private final ReadWriteLock m_stateLock = new ReentrantReadWriteLock();
-
   private OdometryThread m_odometryThread;
 
   private RAROdometry() {
     super("Odometry");
 
-    m_limelight = new Limelight("limelight");
+    m_limelight = new Limelight("limelight", m_stateLock, LimelightType.LL4);
     m_gyro = new AHRS(AHRS.NavXComType.kMXP_SPI, NavXUpdateRate.k200Hz);
 
     m_odometryThread = new OdometryThread(m_stateLock);
@@ -68,8 +66,6 @@ public class RAROdometry extends Subsystem {
         },
         new Pose2d(0, 0, Rotation2d.fromDegrees(0)) // we clarified this works
     );
-
-    // m_visionConstants = new VisionConstants(1,100,0,100);
   }
 
   public static RAROdometry getInstance() {
@@ -144,69 +140,14 @@ public class RAROdometry extends Subsystem {
     setPose(pose);
   }
 
-  Pose2d nullPose = new Pose2d();
-
-  private boolean isPoseZero(PoseEstimate estimate) {
-    return estimate.pose.equals(nullPose);
-  }
-
-  private boolean checkPose(PoseEstimate estimate) {
-    if (estimate == null) {
-      return false;
-    }
-
-    if (isPoseZero(estimate)) {
-      return false;
-    }
-
-    if (estimate.pose.getX() <= 0 || estimate.pose.getX() > RobotConstants.robotConfig.Field.k_length) {
-      return false;
-    }
-
-    if (estimate.pose.getY() <= 0 || estimate.pose.getY() > RobotConstants.robotConfig.Field.k_width) {
-      return false;
-    }
-
-    if (estimate.tagCount <= 0) {
-      return false;
-    }
-
-    if (Math.abs(m_gyro.getRate()) > 720) {
-      return false;
-    }
-
-    return true;
-  }
-
-  // TODO: Add these to Constants when we're done testing them
-  private double xyStdDevCoefficient = 0.005;
-  private double thetaStdDevCoefficient = 0.01;
-  private double stdDevFactor = 0.5; // TODO: Add more!
-  private boolean useVisionRotation = true;
-
-  private void updatePoseWithStdDev(PoseEstimate estimate) {
-    double avgDistance = estimate.avgTagDist;
-    double xyStdDev = xyStdDevCoefficient
-        * Math.pow(avgDistance, 2.0)
-        / estimate.tagCount
-        * stdDevFactor
-        * (DriverStation.isAutonomous() ? m_visionConstants.autoStdDevScale : 1.0);
-
-    double thetaStdDev = useVisionRotation
-        ? thetaStdDevCoefficient
-            * Math.pow(avgDistance, 2.0)
-            / estimate.tagCount
-            * stdDevFactor
-            * (DriverStation.isAutonomous() ? m_visionConstants.autoStdDevScale : 1.0)
-        : Double.POSITIVE_INFINITY;
-
+  public void addLLPose(PoseEstimate estimate, Vector<N3> stdDevs) {
     try {
       m_stateLock.writeLock().lock();
 
       m_poseEstimator.addVisionMeasurement(
           estimate.pose,
           estimate.timestampSeconds,
-          VecBuilder.fill(xyStdDev, xyStdDev, thetaStdDev));
+          stdDevs);
     } finally {
       m_stateLock.writeLock().unlock();
     }
@@ -214,36 +155,18 @@ public class RAROdometry extends Subsystem {
 
   @Override
   public void reset() {
-    // m_odometryThread.stop();
-
-    // try {
-    // m_stateLock.writeLock().lockInterruptibly();
-    // } catch (InterruptedException exception) {
-    // exception.printStackTrace();
-    // }
-
     m_stateLock.writeLock().lock();
-    // m_stateLock.readLock().lock();
 
     resetGyro();
     resetOdometry(new Pose2d());
 
     m_stateLock.writeLock().unlock();
-    // m_stateLock.readLock().unlock();
-
-    // m_odometryThread.start();
   }
 
   @Override
   public void periodic() {
     if (!m_odometryThread.isRunning()) {
       m_odometryThread.start();
-    }
-
-    PoseEstimate estimate = m_limelight.getPoseEstimation();
-
-    if (checkPose(estimate)) {
-      updatePoseWithStdDev(estimate);
     }
 
     logAprilTagData();
